pyproject.toml
[project]
name = "caden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "chz>=0.4.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy<2.3",
    "pandas>=3.0.0",
    "seaborn>=0.13.2",
    "toml>=0.10.2",
    "tqdm>=4.66.3",
    "wandb>=0.24.1",
]

[project.optional-dependencies]
gpu = [
    "inspect-ai>=0.3.174",
    "torch>=2.9.1",
    "vllm==0.15.1",
    "llmcompressor>=0.9.0.1",
]
evals = [
    "backoff>=2.2.1",
    "datasets>=4.0.0",
    "docent-python>=0.1.47a0",
]
